feat [김재성]: 카드리스트 컴포넌트 분리
