+++ components/card/card-list-header.tsx
@@ -1,0 +1,5 @@
+const CardListHeader = () => {
+  return <div>Enter</div>;
+};
+
+export default CardListHeader;
